requirements.txt
aiohttp==1.0.5
async-timeout==2.0.0
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
discord.py==0.16.12
idna==2.6
matplotlib==2.1.0
multidict==3.3.2
nose==1.3.7
numpy==1.13.3
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
websockets==3.4
